requirements.txt
flask==3.0.3
gunicorn==22.0.0
openai>=1.40.0
yt-dlp==2024.8.6
ffmpeg-python==0.2.0
Pillow==10.4.0
opencv-python-headless==4.10.0.84
faiss-cpu==1.8.0.post1
pypdf==4.3.1
python-docx==1.1.2
tiktoken==0.7.0
nltk==3.9.1
pdfkit==1.0.0
numpy==1.26.4